third_party/blink/web_tests/editing/caret/caret-height.html

<!DOCTYPE html>
<html>
<body>
<input id="line1" style="line-height: 1; font-size: 11px;"/>
<input id="line2" style="line-height: 2; font-size: 11px;">
<script src="../../resources/js-test.js"></script>
<script>
    description('Tests whether explicitly set caret style is retained on performing text selection');
    var input1 = document.getElementById('line1');
    var input2 = document.getElementById('line2');
    input1.focus();
    var cursorLine1 = internals.absoluteCaretBounds(input1).height;
    input2.focus();
    var cursorLine2 = internals.absoluteCaretBounds(input2).height;
    shouldBe('cursorLine1', 'cursorLine2');
</script>
</body>
</html>
