third_party/blink/web_tests/editing/caret/caret-height-expected.txt

Tests whether explicitly set caret style is retained on performing text selection

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS cursorLine1 is cursorLine2
PASS successfullyParsed is true

TEST COMPLETE

 
